Add cancel blood drink event
